refactor(importer): rename open_cpap_import → cpap_parser_import; fix event key-lookup bug; add tz and event tests
- git mv open_cpap_import.py → cpap_parser_import.py (and test file)
- Fix stale docstring: parser_validated is True for Lowenstein, False for ResMed
- Fix event key-lookup: localize session_start key in events_by_start so lookup
  matches the localized start_datetime (naive != tz-aware dict key mismatch)
- Add test_run_import_localizes_naive_datetime
- Add test_run_import_events_regrouped
